+++ lib/jobs/geomancer.lua
@@ -704,6 +704,7 @@ return {
     
     -- Action priority order
     priority_order = {
+        'item',
         'heal_aoe',
         'heal_pet',
         'heal',