+++ lib/jobs/paladin.lua
@@ -195,6 +195,7 @@ return {
     
     -- Action priority order
     priority_order = {
+        'item',
         'heal',
         'buff',
         -- recover,