+++ lib/jobs/rune_fencer.lua
@@ -460,6 +460,7 @@ return {
     
     -- Action priority order
     priority_order = {
+        'item',
         'heal',
         'buff',
     },